+++ csharp/src/Ice/InstrumentationI.cs
@@ -25,13 +25,13 @@ namespace ZeroC.Ice
     internal class CollocatedInvocationHelper : MetricsHelper<CollocatedMetrics>
     {
         private static readonly AttributeResolver _attributeResolver = new AttributeResolverI();
-        private readonly int _requestId;
+        private readonly long _requestId;
         private readonly string _id;
         private readonly int _size;
 
         public override void InitMetrics(CollocatedMetrics v) => v.Size += _size;
 
-        internal CollocatedInvocationHelper(ObjectAdapter adapter, int requestId, int size)
+        internal CollocatedInvocationHelper(ObjectAdapter adapter, long requestId, int size)
             : base(_attributeResolver)
         {
             _id = adapter.Name;
@@ -135,7 +135,7 @@ namespace ZeroC.Ice
             return null;
         }
 
-        public IDispatchObserver? GetDispatchObserver(Current current, int requestId, int size)
+        public IDispatchObserver? GetDispatchObserver(Current current, long requestId, int size)
         {
             if (_dispatch.IsEnabled)
             {
@@ -321,14 +321,14 @@ namespace ZeroC.Ice
         private static readonly AttributeResolver _attributeResolver = new AttributeResolverI();
 
         private readonly Current _current;
-        private readonly int _requestId;
+        private readonly long _requestId;
 
         private string? _id;
         private readonly int _size;
 
         public override void InitMetrics(DispatchMetrics v) => v.Size += _size;
 
-        internal DispatchHelper(Current current, int requestId, int size)
+        internal DispatchHelper(Current current, long requestId, int size)
             : base(_attributeResolver)
         {
             _current = current;
@@ -524,7 +524,7 @@ namespace ZeroC.Ice
     internal class InvocationObserver : ObserverWithDelegate<InvocationMetrics, IInvocationObserver>,
         IInvocationObserver
     {
-        public ICollocatedObserver? GetCollocatedObserver(ObjectAdapter adapter, int requestId, int size) =>
+        public ICollocatedObserver? GetCollocatedObserver(ObjectAdapter adapter, long requestId, int size) =>
             GetObserver<CollocatedMetrics, CollocatedObserver, ICollocatedObserver>(
                 "Collocated",
                 new CollocatedInvocationHelper(adapter, requestId, size),
@@ -532,7 +532,7 @@ namespace ZeroC.Ice
 
         public IRemoteObserver? GetRemoteObserver(
             Connection connection,
-            int requestId,
+            long requestId,
             int size) =>
             GetObserver<RemoteMetrics, RemoteObserver, IRemoteObserver>(
                 "Remote",
@@ -643,13 +643,13 @@ namespace ZeroC.Ice
         private static readonly AttributeResolver _attributeResolver = new AttributeResolverI();
 
         private readonly Connection _connection;
-        private readonly int _requestId;
+        private readonly long _requestId;
         private readonly int _size;
         private string? _id;
 
         public override void InitMetrics(RemoteMetrics v) => v.Size += _size;
 
-        internal RemoteInvocationHelper(Connection connection, int requestId, int size)
+        internal RemoteInvocationHelper(Connection connection, long requestId, int size)
             : base(_attributeResolver)
         {
             _connection = connection;
